+++ src/game/interfaces/replay.interface.ts
@@ -1,0 +1,14 @@
+interface IReplay {
+  avatar: string;
+  nick: string;
+  beatmapId: number;
+  score: number;
+  accuracy: number;
+  perfectCount: number;
+  goodCount: number;
+  badCount: number;
+  missCount: number;
+  maxCombo: number;
+}
+
+export { IReplay };